requirements.txt
# Base ----------------------------------------
sqlalchemy==2.0.32
pandas==2.2.2
python-dotenv==0.20.0
psycopg2-binary==2.9.3

numpy>=1.26.4
# opencv-python>=4.1.2
# matplotlib>=3.2.2
# ipyleaflet>=0.14.0
# sympy>=1.10.1
jupyter==1.*

# Machine learning ----------------------------
# scikit-learn
